skip copy of css/js/icons on pre render if is production run (uses symlinks instead)

+++ R/pre_render.R
@@ -4,20 +4,23 @@ input_dirs <- c("css", "js", "icons")
 output_dirs <- c("daily", "monthly", "seasonal") |>
   file.path("index_files")
 
-for (output_dir in output_dirs) {
-  dir.create(output_dir, showWarnings = FALSE)
-  for (input_dir in input_dirs) {
-    if (!dir.exists(input_dir)) {
-      next
+# skip if on production server (uses symlinks instead)
+if (!is_production){
+  for (output_dir in output_dirs) {
+    dir.create(output_dir, showWarnings = FALSE)
+    for (input_dir in input_dirs) {
+      if (!dir.exists(input_dir)) {
+        next
+      }
+      out_dir <- output_dir |> file.path(input_dir)
+      out_dir |> unlink(recursive = TRUE, force = TRUE)
+      file.copy(
+        from = input_dir,
+        to = output_dir,
+        recursive = TRUE,
+        overwrite = TRUE
+      ) |>
+        invisible()
     }
-    out_dir <- output_dir |> file.path(input_dir)
-    out_dir |> unlink(recursive = TRUE, force = TRUE)
-    file.copy(
-      from = input_dir,
-      to = output_dir,
-      recursive = TRUE,
-      overwrite = TRUE
-    ) |>
-      invisible()
   }
 }
